+++ base/exports.jl
@@ -638,6 +638,7 @@ export
 
 # tasks and conditions
     Condition,
+    Event,
     current_task,
     islocked,
     istaskdone,