+++ base/locks-mt.jl
@@ -1,23 +1,18 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
 import .Base: _uv_hook_close, unsafe_convert,
-    lock, trylock, unlock, islocked, wait, notify
+    lock, trylock, unlock, islocked, wait, notify,
+    AbstractLock, GenericCondition, GenericReentrantLock, GenericEvent
 
-export SpinLock, RecursiveSpinLock, Mutex, Event
+export ConditionMT, EventMT, ReentrantLockMT
 
+# Important Note: these low-level primitives exported here
+#   are typically not for general usage
+export SpinLock, RecursiveSpinLock, Mutex
 
 ##########################################
 # Atomic Locks
 ##########################################
-
-"""
-    AbstractLock
-
-Abstract supertype describing types that
-implement the thread-safe synchronization primitives:
-[`lock`](@ref), [`trylock`](@ref), [`unlock`](@ref), and [`islocked`](@ref).
-"""
-abstract type AbstractLock end
 
 # Test-and-test-and-set spin locks are quickest up to about 30ish
 # contending threads. If you have more contention than that, perhaps
@@ -107,6 +102,7 @@ may be held for a considerable length of time.
 """
 const RecursiveSpinLock = RecursiveTatasLock
 
+
 function lock(l::RecursiveTatasLock)
     if l.ownertid[] == threadid()
         l.handle[] += 1
@@ -141,15 +137,34 @@ function trylock(l::RecursiveTatasLock)
 end
 
 function unlock(l::RecursiveTatasLock)
-    @assert(l.ownertid[] == threadid(), "unlock from wrong thread")
-    @assert(l.handle[] != 0, "unlock count must match lock count")
+    l.ownertid[] == threadid() || error("unlock from wrong thread")
+    n = l.handle[]
+    n != 0 || error("unlock count must match lock count")
     if l.handle[] == 1
         l.ownertid[] = 0
         l.handle[] = 0
         ccall(:jl_cpu_wake, Cvoid, ())
     else
-        l.handle[] -= 1
+        l.handle[] = n - 1
     end
+    return
+end
+
+function unlockall(l::RecursiveTatasLock)
+    l.ownertid[] == threadid() || error("unlock from wrong thread")
+    n = l.handle[]
+    n != 0 || error("unlock count must match lock count")
+    l.ownertid[] = 0
+    l.handle[] = 0
+    ccall(:jl_cpu_wake, Cvoid, ())
+    return n
+end
+
+function relockall(l::RecursiveTatasLock, n::Int)
+    lock(l)
+    n1 = l.handle[]
+    l.handle[] = n
+    n1 == 1 || error("concurrency violation detected")
     return
 end
 
@@ -162,13 +177,7 @@ end
 # System Mutexes
 ##########################################
 
-# These are mutexes from libuv. We're doing some error checking (and
-# paying for it in overhead), but regardless, in some situations,
-# passing a bad parameter will cause an abort.
-
-# TODO: how defensive to get, and how to turn it off?
-# TODO: how to catch an abort?
-
+# These are mutexes from libuv.
 const UV_MUTEX_SIZE = ccall(:jl_sizeof_uv_mutex, Cint, ())
 
 """
@@ -205,9 +214,7 @@ function _uv_hook_close(x::Mutex)
 end
 
 function lock(m::Mutex)
-    if m.ownertid == threadid()
-        return
-    end
+    m.ownertid == threadid() && error("concurrency violation detected") # deadlock
     # Temporary solution before we have gc transition support in codegen.
     # This could mess up gc state when we add codegen support.
     gc_state = ccall(:jl_gc_safe_enter, Int8, ())
@@ -218,9 +225,7 @@ function lock(m::Mutex)
 end
 
 function trylock(m::Mutex)
-    if m.ownertid == threadid()
-        return true
-    end
+    m.ownertid == threadid() && error("concurrency violation detected") # deadlock
     r = ccall(:uv_mutex_trylock, Cint, (Ptr{Cvoid},), m)
     if r == 0
         m.ownertid = threadid()
@@ -229,7 +234,7 @@ function trylock(m::Mutex)
 end
 
 function unlock(m::Mutex)
-    @assert(m.ownertid == threadid(), "unlock from wrong thread")
+    m.ownertid == threadid() || error("concurrency violation detected")
     m.ownertid = 0
     ccall(:uv_mutex_unlock, Cvoid, (Ptr{Cvoid},), m)
     return
@@ -240,55 +245,33 @@ function islocked(m::Mutex)
 end
 
 """
-    Event()
+    ReentrantLockMT()
 
-Create a level-triggered event source. Tasks that call [`wait`](@ref) on an
-`Event` are suspended and queued until `notify` is called on the `Event`.
-After `notify` is called, the `Event` remains in a signaled state and
-tasks will no longer block when waiting for it.
+A thread-safe version of [`ReentrantLock`](@ref).
 """
-mutable struct Event
-    lock::Mutex
-    q::Vector{Task}
-    set::Bool
-    # TODO: use a Condition with its paired lock
-    Event() = new(Mutex(), Task[], false)
-end
+const ReentrantLockMT = GenericReentrantLock{TatasLock}
 
-function wait(e::Event)
-    e.set && return
-    lock(e.lock)
-    while !e.set
-        ct = current_task()
-        push!(e.q, ct)
-        unlock(e.lock)
-        try
-            wait()
-        catch
-            filter!(x->x!==ct, e.q)
-            rethrow()
-        end
-        lock(e.lock)
-    end
-    unlock(e.lock)
-    return nothing
-end
+"""
+    ConditionMT([lock-mt])
 
-function notify(e::Event)
-    lock(e.lock)
-    if !e.set
-        e.set = true
-        for t in e.q
-            schedule(t)
-        end
-        empty!(e.q)
-    end
-    unlock(e.lock)
-    return nothing
-end
+A thread-safe version of [`Condition`](@ref).
+"""
+const ConditionMT = GenericCondition{ReentrantLockMT}
 
-# TODO: decide what to call this
-#function clear(e::Event)
-#    e.set = false
-#    return nothing
-#end
+"""
+    EventMT()
+
+A thread-safe version of [`Event`](@ref).
+"""
+const EventMT = GenericEvent{ReentrantLockMT}
+
+"""
+Special note for [`Threads.ConditionMT`](@ref):
+
+The caller must be holding the [`lock`](@ref) that owns `c` before calling this method.
+The calling task will be blocked until some other task wakes it,
+usually by calling [`notify`](@ref)` on the same ConditionMT object.
+The lock will be atomically released when blocking (even if it was locked recursively),
+and will be reacquired before returning.
+"""
+wait(c::ConditionMT)