+++ base/threads.jl
@@ -7,7 +7,8 @@ module Threads
 
 include("threadingconstructs.jl")
 include("atomics.jl")
-include("locks.jl")
+include("locks-mt.jl")
+
 
 """
     resize_nthreads!(A, copyvalue=A[1])