+++ test/threads.jl
@@ -96,7 +96,7 @@ end
 # Check if the recursive lock can be locked and unlocked correctly.
 let critical = RecursiveSpinLock()
     @test !islocked(critical)
-    @test_throws AssertionError unlock(critical)
+    @test_throws ErrorException("unlock from wrong thread") unlock(critical)
     @test lock(critical) === nothing
     @test islocked(critical)
     @test lock(critical) === nothing
@@ -108,12 +108,12 @@ let critical = RecursiveSpinLock()
     @test islocked(critical)
     @test unlock(critical) === nothing
     @test !islocked(critical)
-    @test_throws AssertionError unlock(critical)
+    @test_throws ErrorException("unlock from wrong thread") unlock(critical)
     @test trylock(critical) == true
     @test islocked(critical)
     @test unlock(critical) === nothing
     @test !islocked(critical)
-    @test_throws AssertionError unlock(critical)
+    @test_throws ErrorException("unlock from wrong thread") unlock(critical)
     @test !islocked(critical)
 end
 
@@ -504,9 +504,10 @@ function test_thread_too_few_iters()
 end
 test_thread_too_few_iters()
 
-let e = Event()
+let e = Event(), started = Event()
     done = false
-    t = @async (wait(e); done = true)
+    t = @async (notify(started); wait(e); done = true)
+    wait(started)
     sleep(0.1)
     @test done == false
     notify(e)